Move the page <-> topic inline to taxonomy
